Finally fix the issue with displaying movie details When viewing page on Desktop screen.
