+++ src/components/normalCarousel.js
@@ -15,9 +15,8 @@ const NormalCarousel = (props) => {
   const [responsive, setResponsive] = useState(6);
   const itemsPerPage = responsive;
   const totalPages = Math.ceil(data.length / itemsPerPage);
-  const [singleMovie, setSingleMovie] = useState(null);
   const [movieDetails, setMovieDetails] = useState(false);
-  const months = ["January", "February", "March", "April", "May", "June"];
+
 
   const handleClickPrev = () => {
     setCurrentPage((prevPage) => prevPage - 1);
@@ -35,82 +34,12 @@ const NormalCarousel = (props) => {
     }
   }, [width]);
 
-  const handleCoverClick = () => {
-    console.log(movieDetails);
-    setMovieDetails(false);
-  };
-
-  const handleVideoClick = (e) => {
-    // router.push({
-    //   pathname: `/videos/${id}`,
-    //   query: { videoUrl },
-
-    // })
-    const filterMovie = data.filter((item, index) => index == e.target.id);
-    console.log(e.target);
-    const randomIndex = Math.floor(Math.random() * months.length);
-    const randomDay = Math.floor(Math.random() * 30);
-
-    setSingleMovie(
-      <div className={styles.singleMovie}>
-        <div className={styles.imageBox}>
-          <img src={filterMovie[0].image}></img>
-        </div>
-        <div className={styles.descriptionBox}>
-          <div className={styles.mainInfo}>
-            <div className={styles.grayText}>
-              Duration: {filterMovie[0].duration}m
-            </div>
-            <div>
-              <Image
-                src="/tv-outline.svg"
-                alt="Movie Quality"
-                width={20}
-                height={20}
-              ></Image>
-              <Image
-                src="/subtitles.svg"
-                alt="Subtitle icon"
-                width={20}
-                height={20}
-              ></Image>
-            </div>
-
-            <div className={styles.LastDayWatch}>
-              Last Days to watch on Netflix: {months[randomIndex]} {randomDay}
-            </div>
-            <text>
-              {"Lorem Ipsum is simply dummy text of the printing and typesetting industry. " +
-                "Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, " +
-                "when an unknown printer took a galley of type and scrambled it to make a type " +
-                "specimen book. It has survived not only five centuries, but also the leap into " +
-                "electronic typesetting, remaining essentially unchanged. It was popularised in " +
-                "the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, " +
-                "and more recently with desktop publishing software like Aldus PageMaker " +
-                "including versions of Lorem Ipsum."}
-            </text>
-            <h1>{filterMovie[0].video_files.quality}</h1>
-
-            <div>{filterMovie[0].video_files.fps}</div>
-            <div>{filterMovie[0].video_files.file_type}</div>
-          </div>
-          <div className={styles.secondInfo}>
-            {filterMovie[0].user.name}
-            <div>
-              <p>Cast:</p> {filterMovie[0].user.name} {filterMovie[0].user.name}
-              <p>Genre:</p> {filterMovie[0].user.name}{" "}
-              {filterMovie[0].user.name}
-            </div>
-          </div>
-        </div>
-        <h1>More like this</h1>
-      </div>
-    );
-    if (!movieDetails) {
-      setMovieDetails(true);
-      console.log(itemsPerPage);
-    }
-  };
+  //This function sets movie id in Media component
+  //so then that one can pass it to OverlayPage component
+  //in case user opens the movie for more details
+  const preAssignMovieNumber = (e) =>{
+    props.filteredMovie(e.target.id)
+  }
 
   const startIndex = (currentPage - 1) * itemsPerPage;
   const endIndex = startIndex + itemsPerPage;
@@ -129,7 +58,8 @@ const NormalCarousel = (props) => {
               backgroundImage: `url(${movie.image})`,
               position: "relative",
             }}
-            onClick={handleVideoClick}
+            onClick={props.movieDetailsToggle}
+            onMouseEnter={preAssignMovieNumber}
           >
             <VideoThumbnail
               movieDetails={movieDetails}
@@ -137,13 +67,6 @@ const NormalCarousel = (props) => {
               thumbnailUrl={movie.image}
               videoUrl={movie.video_files.link}
             />
-            {movieDetails && (
-              <MovieDetails
-                id={index}
-                singleMovie={singleMovie}
-                toggle={handleCoverClick}
-              />
-            )}
           </div>
         ))}
       </div>
